IonMonkey: MIPS64: Fix build warning in MacroAssembler::Pop.

+++ js/src/jit/mips64/MacroAssembler-mips64.cpp
@@ -4084,7 +4084,7 @@ void
 MacroAssembler::Pop(Register reg)
 {
     ma_pop(reg);
-    adjustFrame(-sizeof(intptr_t));
+    adjustFrame((int32_t)-sizeof(intptr_t));
 }
 
 void
